+++ lib/widget/event_row_card.dart
